feat: 5주차 1단계 피드백 반영
